demo api module refactor the code
